+++ src/main/java/de/alpharogroup/swing/help/HelpFrame.java
@@ -44,6 +44,8 @@ import de.alpharogroup.layout.DisposeWindow;
 public class HelpFrame extends JFrame
 {
 
+	private static final long serialVersionUID = 1L;
+
 	private static final String DEFAULT_BUTTONLABEL_CLOSE = "Close";
 
 	private static final String DEFAULT_LABEL_HELP = "License";